requirements.txt
aiosqlite==0.18.0
hikari==2.0.0.dev116
hikari-lightbulb==2.3.1
python-dotenv==0.21.1
uvloop==0.17.0
hikari-miru==3.0.1